erm
I lost my c drive 3 days ago which had the old db manager on it
idk why I never pushed it to github, but its lost so I had to rewrite all of this.

It is kind of a blessing but I've rewritten this to be more usable and error proof. Also I updated danser so now it records videos on the latest AMD drivers.

Next step is to get this working in a container so I dont have to upload everything from my home pc. ALthough I do not think that will be easy.